+++ src/main/java/gov/dhs/nppd/humanreview/controller/HumanreviewApiController.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import javax.validation.Valid;
 
 import org.openapitools.api.HumanreviewApi;
-import org.openapitools.model.HumanReviewItem;
-import org.openapitools.model.ListOfHumanReviewItems;
+import org.openapitools.model.humanreview.HumanReviewItem;
+import org.openapitools.model.humanreview.ListOfHumanReviewItems;
 import org.openapitools.repository.humanreview.HumanreviewRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
