+++ src/main/java/org/openapitools/model/humanreview/APIEndpointParameters.java
@@ -1,0 +1,130 @@
+package org.openapitools.model.humanreview;
+
+import java.util.Objects;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import javax.validation.Valid;
+import javax.validation.constraints.*;
+
+/**
+ * APIEndpointParameters
+ */
+@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2018-10-11T09:52:13.189-04:00[America/New_York]")
+
+public class APIEndpointParameters   {
+  @JsonProperty("type")
+  private String type = null;
+
+  @JsonProperty("description")
+  private String description = null;
+
+  @JsonProperty("required")
+  private Boolean required = null;
+
+  public APIEndpointParameters type(String type) {
+    this.type = type;
+    return this;
+  }
+
+  /**
+   * Get type
+   * @return type
+  **/
+  @ApiModelProperty(required = true, value = "")
+  @NotNull
+
+
+  public String getType() {
+    return type;
+  }
+
+  public void setType(String type) {
+    this.type = type;
+  }
+
+  public APIEndpointParameters description(String description) {
+    this.description = description;
+    return this;
+  }
+
+  /**
+   * Get description
+   * @return description
+  **/
+  @ApiModelProperty(value = "")
+
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  public APIEndpointParameters required(Boolean required) {
+    this.required = required;
+    return this;
+  }
+
+  /**
+   * Get required
+   * @return required
+  **/
+  @ApiModelProperty(value = "")
+
+
+  public Boolean getRequired() {
+    return required;
+  }
+
+  public void setRequired(Boolean required) {
+    this.required = required;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    APIEndpointParameters apIEndpointParameters = (APIEndpointParameters) o;
+    return Objects.equals(this.type, apIEndpointParameters.type) &&
+        Objects.equals(this.description, apIEndpointParameters.description) &&
+        Objects.equals(this.required, apIEndpointParameters.required);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(type, description, required);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class APIEndpointParameters {\n");
+    
+    sb.append("    type: ").append(toIndentedString(type)).append("\n");
+    sb.append("    description: ").append(toIndentedString(description)).append("\n");
+    sb.append("    required: ").append(toIndentedString(required)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
+}
+
