+++ src/test/java/gov/dhs/nppd/humanreview/service/HumanreviewApiServiceTest.java
@@ -13,9 +13,9 @@ import javax.persistence.EntityManager;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
-import org.openapitools.model.HumanReviewItem;
-import org.openapitools.model.JsonData;
-import org.openapitools.model.ListOfHumanReviewItems;
+import org.openapitools.model.humanreview.HumanReviewItem;
+import org.openapitools.model.humanreview.JsonData;
+import org.openapitools.model.humanreview.ListOfHumanReviewItems;
 import org.openapitools.repository.humanreview.HumanreviewRepository;
 import org.openapitools.repository.humanreview.JsonDataRepository;
 import org.springframework.context.annotation.ComponentScan;
